+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
 ### New features and enhancements
 
 - New API call `getJobsByParameters` to allow filtering jobs by status.
+- New API call `dataSetsMatchingPattern` to allow filtering datasets via a pattern.
 - Added `findEquivalentNode` function to IZoweTree to find a corresponding favorited/non-favorited node.
 - Updated `IZoweTree`: changed `IZoweNodeType -> IZoweTreeNode` to prevent incompatibility w/ custom/future Zowe node types